+++ server/listener/requesthandler/requesthandler.go
@@ -21,6 +21,7 @@ import (
 	"context"
 	"reflect"
 
+	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/golang/protobuf/proto"
 	"github.com/nymtech/nym/common/comm/commands"
 	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
@@ -186,5 +187,21 @@ func ResolveSpendCredentialRequestHandler(ctx context.Context, resCh <-chan *com
 	return &commands.SpendCredentialResponse{
 		WasSuccessful: status,
 		Status:        protoStatus,
+	}
+}
+
+func ResolveFaucetTransferRequest(ctx context.Context, resCh <-chan *commands.Response) proto.Message {
+	data, protoStatus := waitUntilResolved(ctx, resCh)
+	erc20TxHash := make([]byte, ethcommon.HashLength)
+	etherTxHash := make([]byte, ethcommon.HashLength)
+	if data != nil {
+		hashes := data.([]byte)
+		i := copy(erc20TxHash, hashes)
+		copy(etherTxHash, hashes[i:])
+	}
+	return &commands.FaucetTransferResponse{
+		Erc20TxHash: erc20TxHash,
+		EtherTxHash: etherTxHash,
+		Status:      protoStatus,
 	}
 }